Update template int -> long

+++ Common/IProblem.cs
@@ -4,7 +4,7 @@ namespace AOC
     {
         public void Parse(string[] input);
 
-        public int PartOne();
-        public int PartTwo();
+        public long PartOne();
+        public long PartTwo();
     }
 }
